requirements.txt
mysql-connector==2.2.9
pkg-resources==0.0.0
